feat: 앨범 조회 페이지 구현
